Adds ability for custom cookie file lookup for daemon(core) auth.
The .cookie is a daemon specific file for genearting random authentication details that a node can use to authenticate rpc connection details. According to the spec, the name of the file defaults to '.cookie', but an option does exist to provide an alternative name for this file.
This change simply allows for an alternative daemon authentication cookie filename to be specified, in case it exists.