Fixes expired listing items being deleted where the listing item was part of the user's buyflow
Currently,  a ListingItem with an expiry date is selected for removal, the comments and images for the item are deleted, and then the item itself. Except that if the item is attached to a bid/order, then the attempt to remove the listing item fails with a database errror. However, the images and comments for the listing item have since been deleted.
This change prevents a listing item from being selected for deletion if the item currently has any associated bids.